requirements.txt
behave==1.2.6
behave-django==1.3.0
Counter==1.0.0
Django==2.2.1
numpy==1.16.3
pandas==0.24.2
parse==1.12.0
parse-type==0.4.2
PyHamcrest==1.9.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
